Takes over notif prefs
