+++ code/modules/events/wizard/ghost.dm
@@ -24,7 +24,6 @@
 	description = "Ghosts become visible and gain the power of possession."
 
 /datum/round_event/wizard/possession/start()
-	for(var/mob/dead/observer/G in GLOB.player_list)
-		add_verb(G, /mob/dead/observer/verb/boo)
-		add_verb(G, /mob/dead/observer/verb/possess)
-		to_chat(G, "You suddenly feel a welling of new spooky powers...")
+	for(var/mob/dead/observer/ghost_player in GLOB.player_list)
+		ghost_player.fun_verbs = TRUE
+		to_chat(ghost_player, span_hypnophrase("You suddenly feel a welling of new spooky powers..."))
